Solve Part 1 of Problem 2
This solution assumes the functions have that very specific form in the first three lines, but is otherwise fairly general. We computed the median by sorting and taking the middle, when there are O(N) solutions to the problem of finding the median. On 101 numbers, though, there's no real reason to worry about the performance of sorting.