Use `select_nth_unstable()` instead of `sorted()` to get median
I had just sorted the (not very long list) to get the median, but I knew there were $O(N)$ algorithms for finding the medians. I did a little looking into that, and once I realized that what I was looking for were _selection_ algorithms, I remembered that Rust provides those!

This switches from sorting and finding the middle value to calling `.select_nth_unstable()` to do that work, and everything is swell.